+++ client/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed TOO_MANY_REDIRECTS error after terminate session from other device.
+- Fixed Total Rest view metrics.
 
 ## [0.5.0] - 2026-04-04